project/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==10.0.0
importlib-metadata==4.10.0
Markdown==3.3.6
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
zipp==3.7.0
